cleanup part 1
