+++ docs/api-intro.md
@@ -2,7 +2,7 @@
 sidebar_position: 5
 ---
 
-import { AuthButtonWrapper } from '@site/src/components/AuthButton';
+import { AuthButtonWrapper } from '@site/src/components/AuthButtonWrapper';
 
 <div style={{ textAlign: 'right' }}>
   <AuthButtonWrapper />
